docs(inferencex): update plan and README for flat layout and tiered gates
