+++ cvs/tests/inference/inferencex_atom/conftest.py
@@ -10,6 +10,13 @@ import pytest
 
 from cvs.core.orchestrators.factory import OrchestratorConfig, OrchestratorFactory
 from cvs.lib import globals
+from cvs.lib.inference.inference_suite_lifecycle import (
+    InferenceLifecycle,
+    attach_lifecycle_html_table,
+    html_metric_table_header,
+    html_metric_table_row,
+    sort_lifecycle_items,
+)
 from cvs.lib.inference.utils.inferencex_atom_config_loader import (
     load_variant,
     orchestrator_container_from_variant,
@@ -18,25 +25,25 @@ from cvs.lib.utils_lib import resolve_cluster_config_placeholders
 
 log = globals.log
 
+LIFECYCLE_RANK = {
+    "test_launch_container": 0,
+    "test_setup_sshd": 1,
+    "test_model_fetch": 2,
+    "test_inferencex_atom_inference": 3,
+    "test_cell_metrics": 4,
+    "test_print_results_table": 5,
+    "test_teardown": 6,
+}
+
 
 def _deep_merge(base, override):
-    """Recursively merge `override` onto `base` (dicts merged key-wise, scalars/lists replaced)."""
+    """Recursively merge ``override`` onto ``base`` (dicts merged key-wise; scalars/lists replaced)."""
     if not (isinstance(base, dict) and isinstance(override, dict)):
         return override
     out = dict(base)
     for k, v in override.items():
         out[k] = _deep_merge(base[k], v) if k in base else v
     return out
-
-
-class _Lifecycle:
-    def __init__(self):
-        self.failed = False
-        self.torn_down = False
-        self.report = {}
-
-    def record(self, nodeid, label, value, unit="s"):
-        self.report.setdefault(nodeid, []).append((label, value, unit))
 
 
 @pytest.fixture(scope="module")
@@ -59,7 +66,7 @@ def variant_config(pytestconfig, cluster_dict):
 
 @pytest.fixture(scope="module")
 def lifecycle():
-    return _Lifecycle()
+    return InferenceLifecycle()
 
 
 @pytest.fixture(scope="module")
@@ -100,59 +107,18 @@ def inf_res_dict():
 
 
 def pytest_collection_modifyitems(items):
-    rank = {
-        "test_launch_container": 0,
-        "test_setup_sshd": 1,
-        "test_model_fetch": 2,
-        "test_inferencex_atom_inference": 3,
-        "test_cell_metrics": 4,
-        "test_print_results_table": 5,
-        "test_teardown": 6,
-    }
-    items.sort(key=lambda it: rank.get(it.originalname or it.name.split("[")[0], 99))
+    sort_lifecycle_items(items, LIFECYCLE_RANK)
 
 
 @pytest.hookimpl(hookwrapper=True)
 def pytest_runtest_makereport(item, call):
     outcome = yield
-    report = outcome.get_result()
-    if report.when != "call":
-        return
-    lc = item.funcargs.get("lifecycle")
-    rows = getattr(lc, "report", {}).get(item.nodeid) if lc else None
-    if not rows:
-        return
-    try:
-        import pytest_html
-    except ImportError:
-        return
-    body = "".join(
-        f"<tr><td>{label}</td><td>{value:.1f}</td><td>{unit}</td></tr>"
-        for label, value, unit in rows
-    )
-    html = f"<table><tr><th>stage</th><th>value</th><th>unit</th></tr>{body}</table>"
-    extras = getattr(report, "extras", [])
-    extras.append(pytest_html.extras.html(html))
-    report.extras = extras
+    attach_lifecycle_html_table(item, outcome.get_result())
 
 
 def pytest_html_results_table_header(cells):
-    cells.insert(-1, "<th>Value</th>")
-    cells.insert(-1, "<th>Unit</th>")
+    html_metric_table_header(cells)
 
 
 def pytest_html_results_table_row(report, cells):
-    props = dict(report.user_properties)
-    has = "metric_value" in props
-    val = props.get("metric_value")
-    unit = props.get("metric_unit", "") if has else ""
-    if not has:
-        shown = ""
-    elif val is None:
-        shown = "-"
-    elif isinstance(val, float):
-        shown = f"{val:.3f}"
-    else:
-        shown = str(val)
-    cells.insert(-1, f"<td>{shown}</td>")
-    cells.insert(-1, f"<td>{unit}</td>")
+    html_metric_table_row(report, cells)
